some things are still not right
