--- after/plugin/autopairs.lua
@@ -1,35 +1,0 @@
-local status_ok, autopairs = pcall(require, "nvim-autopairs")
-if not status_ok then
-  return
-end
-
-autopairs.setup({
-  check_ts = true,
-  ts_config = {
-    lua = {'string'},
-    javascript = {'template_string'},
-    java = false,
-  },
-  fast_wrap = {
-    map = '<M-e>',
-    chars = { '{', '[', '(', '"', "'" },
-    pattern = [=[[%'%"%)%>%]%)%}%,]]=],
-    end_key = '$',
-    keys = 'qwertyuiopzxcvbnmasdfghjkl',
-    check_comma = true,
-    highlight = 'Search',
-    highlight_grey = 'Comment'
-  },
-  disable_filetype = { "TelescopePrompt", "vim" },
-})
-
--- Setup integration with cmp if available
-local cmp_status_ok, cmp = pcall(require, "cmp")
-if cmp_status_ok then
-  -- Integration with cmp
-  local cmp_autopairs = require('nvim-autopairs.completion.cmp')
-  cmp.event:on(
-    'confirm_done',
-    cmp_autopairs.on_confirm_done()
-  )
-end
