--- after/plugin/comment.lua
@@ -1,7 +1,0 @@
-require('Comment').setup({
-    sticky = true,
-    mappings = {
-        basic = true,
-        extra = true,
-    },
-})
