--- after/plugin/noneckpain.lua
@@ -1,32 +1,0 @@
-require("no-neck-pain").setup({
-    width = 120,
-    minSideBufferWidth = 10,
-    autocmds = {
-        enableOnVimEnter = false,
-    },
-    mappings = {
-        enabled = true,
-    },
-    buffers = {
-        scratchPad = {
-            enabled = true,
-            location = "~/Git/ScratchPadFiles/"
-        },
-        bo = {
-            filetype = "md",
-        },
-        -- not working for some reason!?
-        colors = {
-            backgroundColor = "#525252",
-            right = {
-                backgroundColor = "tokyonight-moon",
-            },
-            left = {
-                backgroundColor = "tokyonight-moon",
-            },
-        },
-        wo = {
-            fillchars = "eob: ",
-        }
-    }
-})
