--- after/plugin/todo.lua
@@ -1,3 +1,0 @@
-local todo = require("todo-comments")
-todo.setup()
-
