+++ init.lua
@@ -1,1 +1,2 @@
 require("lio")
+require("lio.lazy")
