+++ lua/lio/init.lua
@@ -1,2 +1,4 @@
 require("lio.remap")
 require("lio.set")
+
+require("lio.floatterminal")
